Move _const methods from CallExt to Call
