+++ crates/nu-engine/src/call_ext.rs
@@ -1,7 +1,6 @@
 use nu_protocol::{
-    ast::{Call, Expression},
-    engine::{EngineState, Stack, StateWorkingSet},
-    eval_const::eval_constant,
+    ast::Call,
+    engine::{EngineState, Stack},
     FromValue, ShellError, Value,
 };
 
@@ -16,24 +15,10 @@ pub trait CallExt {
         flag_name: &str,
     ) -> Result<bool, ShellError>;
 
-    /// Check if a boolean flag is set (i.e. `--bool` or `--bool=true`)
-    /// evaluating the expression after = as a constant command
-    fn has_flag_const(
-        &self,
-        working_set: &StateWorkingSet,
-        flag_name: &str,
-    ) -> Result<bool, ShellError>;
-
     fn get_flag<T: FromValue>(
         &self,
         engine_state: &EngineState,
         stack: &mut Stack,
-        name: &str,
-    ) -> Result<Option<T>, ShellError>;
-
-    fn get_flag_const<T: FromValue>(
-        &self,
-        working_set: &StateWorkingSet,
         name: &str,
     ) -> Result<Option<T>, ShellError>;
 
@@ -43,16 +28,6 @@ pub trait CallExt {
         stack: &mut Stack,
         starting_pos: usize,
     ) -> Result<Vec<T>, ShellError>;
-
-    fn rest_const<T: FromValue>(
-        &self,
-        working_set: &StateWorkingSet,
-        starting_pos: usize,
-    ) -> Result<Vec<T>, ShellError>;
-
-    fn rest_iter_flattened<F>(&self, start: usize, eval: F) -> Result<Vec<Value>, ShellError>
-    where
-        F: FnMut(&Expression) -> Result<Value, ShellError>;
 
     fn opt<T: FromValue>(
         &self,
@@ -65,12 +40,6 @@ pub trait CallExt {
         &self,
         engine_state: &EngineState,
         stack: &mut Stack,
-        pos: usize,
-    ) -> Result<T, ShellError>;
-
-    fn req_const<T: FromValue>(
-        &self,
-        working_set: &StateWorkingSet,
         pos: usize,
     ) -> Result<T, ShellError>;
 
@@ -112,34 +81,6 @@ impl CallExt for Call {
         Ok(false)
     }
 
-    fn has_flag_const(
-        &self,
-        working_set: &StateWorkingSet,
-        flag_name: &str,
-    ) -> Result<bool, ShellError> {
-        for name in self.named_iter() {
-            if flag_name == name.0.item {
-                return if let Some(expr) = &name.2 {
-                    // Check --flag=false
-                    let result = eval_constant(working_set, expr)?;
-                    match result {
-                        Value::Bool { val, .. } => Ok(val),
-                        _ => Err(ShellError::CantConvert {
-                            to_type: "bool".into(),
-                            from_type: result.get_type().to_string(),
-                            span: result.span(),
-                            help: Some("".into()),
-                        }),
-                    }
-                } else {
-                    Ok(true)
-                };
-            }
-        }
-
-        Ok(false)
-    }
-
     fn get_flag<T: FromValue>(
         &self,
         engine_state: &EngineState,
@@ -148,19 +89,6 @@ impl CallExt for Call {
     ) -> Result<Option<T>, ShellError> {
         if let Some(expr) = self.get_flag_expr(name) {
             let result = eval_expression(engine_state, stack, expr)?;
-            FromValue::from_value(result).map(Some)
-        } else {
-            Ok(None)
-        }
-    }
-
-    fn get_flag_const<T: FromValue>(
-        &self,
-        working_set: &StateWorkingSet,
-        name: &str,
-    ) -> Result<Option<T>, ShellError> {
-        if let Some(expr) = self.get_flag_expr(name) {
-            let result = eval_constant(working_set, expr)?;
             FromValue::from_value(result).map(Some)
         } else {
             Ok(None)
@@ -179,43 +107,6 @@ impl CallExt for Call {
             eval_expression(engine_state, stack, expr)
         })? {
             output.push(FromValue::from_value(result)?);
-        }
-
-        Ok(output)
-    }
-
-    fn rest_const<T: FromValue>(
-        &self,
-        working_set: &StateWorkingSet,
-        starting_pos: usize,
-    ) -> Result<Vec<T>, ShellError> {
-        let mut output = vec![];
-
-        for result in
-            self.rest_iter_flattened(starting_pos, |expr| eval_constant(working_set, expr))?
-        {
-            output.push(FromValue::from_value(result)?);
-        }
-
-        Ok(output)
-    }
-
-    fn rest_iter_flattened<F>(&self, start: usize, mut eval: F) -> Result<Vec<Value>, ShellError>
-    where
-        F: FnMut(&Expression) -> Result<Value, ShellError>,
-    {
-        let mut output = Vec::new();
-
-        for (expr, spread) in self.rest_iter(start) {
-            let result = eval(expr)?;
-            if spread {
-                match result {
-                    Value::List { mut vals, .. } => output.append(&mut vals),
-                    _ => return Err(ShellError::CannotSpreadAsList { span: expr.span }),
-                }
-            } else {
-                output.push(result);
-            }
         }
 
         Ok(output)
@@ -243,24 +134,6 @@ impl CallExt for Call {
     ) -> Result<T, ShellError> {
         if let Some(expr) = self.positional_nth(pos) {
             let result = eval_expression(engine_state, stack, expr)?;
-            FromValue::from_value(result)
-        } else if self.positional_len() == 0 {
-            Err(ShellError::AccessEmptyContent { span: self.head })
-        } else {
-            Err(ShellError::AccessBeyondEnd {
-                max_idx: self.positional_len() - 1,
-                span: self.head,
-            })
-        }
-    }
-
-    fn req_const<T: FromValue>(
-        &self,
-        working_set: &StateWorkingSet,
-        pos: usize,
-    ) -> Result<T, ShellError> {
-        if let Some(expr) = self.positional_nth(pos) {
-            let result = eval_constant(working_set, expr)?;
             FromValue::from_value(result)
         } else if self.positional_len() == 0 {
             Err(ShellError::AccessEmptyContent { span: self.head })
