+++ crates/nu-plugin/src/protocol/evaluated_call.rs
@@ -1,4 +1,4 @@
-use nu_engine::{eval_expression, CallExt};
+use nu_engine::eval_expression;
 use nu_protocol::{
     ast::Call,
     engine::{EngineState, Stack},
